moves AccountBalanceInterface and related types to backend api types
