+++ @shared/api/types/backend-api.ts
@@ -1,6 +1,15 @@
+import { AssetBalance, NativeBalance, TokenBalance } from "./types";
+
 export interface BalanceMap {
-  [key: string]: unknown; // TODO
-  native: unknown; // TODO
+  [key: string]: AssetBalance | NativeBalance | TokenBalance;
+  native: NativeBalance;
 }
 
 export type Balances = BalanceMap | null;
+
+export interface AccountBalancesInterface {
+  balances: Balances;
+  isFunded: boolean | null;
+  subentryCount: number;
+  error?: { horizon: any; soroban: any };
+}
